requirements.txt
aiofiles==25.1.0
aiosqlite==0.22.0
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
appnope==0.1.4
asttokens==3.0.1
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cobble==0.1.4
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.1
google_search_results==2.4.2
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
jupyter_client==8.7.0
jupyter_core==5.9.1
lxml==6.0.2
mammoth==1.11.0
markdown-it-py==4.0.0
markdownify==1.2.2
matplotlib-inline==0.2.1
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathvalidate==3.3.1
pdfminer.six==20251107
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynput==1.8.1
pyobjc-core==12.1
pyobjc-framework-ApplicationServices==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreText==12.1
pyobjc-framework-Quartz==12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
pyzmq==27.1.0
questionary==2.1.1
regex==2025.11.3
requests==2.32.5
rich==14.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
stack-data==0.6.3
tiktoken==0.12.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.14
xlsxwriter==3.2.9
youtube-transcript-api==1.2.3
